+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `--convergencesets` for showing convergence sets definitions.
 - `--cell` and `--cell-registers` for showing cell registers definitions.
+- `--parameters` for showing input / output parameters.
 
 ## [v0.5.1] - 2026-09-01